Refactor files to correct format
