--- src/backend/api_components/middleware.mjs
@@ -1,1 +1,0 @@
-// TODO: CREATE MIDDLEWARE TO VERIFY JWTs.
